// src/interfaces/CadastroProfissionalInterface.ts
export interface CadastroProfissionalInterface {
    id: number;
    nome: string;
    celular: string;
    email: string;
    cpf: number;
    dataNascimento: string;
    cidade: string;
    estado: string;
    pais: string;
    rua: string;
    numero: string;
    bairro: string;
    cep: number;
    complemento: string;
    senha: string;
    salario: number;
}